feat: reduce the amount of timer firing
This brings the write coalescing logic in-line with mplex. Instead of proving a
"we will delay by at most 100us" guarantee, we will send if:

1. We haven't seen any new data for the coalesce delay (100us).
2. We have at least 512 bytes.

From a real-time perspective, this is worse as a slow system might take a while
to actually send the message. However, on most systems, the maximum delay should
be `100us * 512 = 52ms`. Really, given that a yamux message frame is 12
bytes (GAH!), this is more like `(100us * 512 / 13) ~= 4ms`.