fix: fix space accounting in the receive buffer
First, I've removed the atomics because they don't really give us much and are a
bit misleading. We only use them to access the length taking a lock, but we
can't actually _set_ them without holding the lock.

Second, the receive buffer logic wasn't quite correct. The window size is
`len+cap`, not `len+cap+pending`.

The receive buffer looks like:

    |            receive window (10)    |
    |     data      | empty space       |
    |< len (5)     >|< cap (5)         >|
                    |< pending (4) ->|

As data is read, the buffer gets updated like so:

       |         receive window (8)     |
       |     data   | empty space       |
       |< len (3)  >|< cap (5)         >|
                    |< pending (4) ->|

It can then grow as follows (given a "max" of 10):

       |         receive window (10)       |
       |     data   | empty space          |
       |< len (3)  >|< cap (7)            >|
                    |< pending (4) ->|